+++ NEWS.md
@@ -1,3 +1,5 @@
+# conflicted (development version)
+
 # conflicted 1.0.4
 
 * Fixes for dev rlang